fix(daemon): address PR #123 review: data race, types, date bounds, retention pointer
- Fix 1: Add sync.Mutex to fakeStore in recorder_test.go; expose count/at/setFailNext helpers to eliminate data race between recorder goroutine and test goroutine
- Fix 2: Change PRNumber from float64 to int in recordIssueImplemented; update test assertion from float64(99) to 99
- Fix 3: Switch ActivityQuery.To to exclusive upper bound (ts < To) in store; update handlers.go to pass start-of-next-day instead of 23:59:59
- Fix 4: Move activity purge ticker inside if *cfg.ActivityLog.Enabled block so it only runs when recording is enabled
- Fix 5: Make ActivityLogConfig.RetentionDays a *int to distinguish absent (nil→default 90) from explicit zero (no-op); add ptrIntOr helper; update config, store, main, tests; add TestActivityLogConfig_ExplicitZeroRetentionIsKept
- Fix 6: Drop redundant != nil check on cfg.ActivityLog.Enabled (applyDefaults always sets it)

+++ daemon/internal/activity/recorder.go
@@ -173,7 +173,7 @@ func (r *Recorder) recordIssueImplemented(ev sse.Event) error {
 		IssueNumber int    `json:"issue_number"`
 		IssueTitle  string `json:"issue_title"`
 		CLIUsed     string `json:"cli_used"`
-		PRNumber    float64 `json:"pr_number"`
+		PRNumber    int    `json:"pr_number"`
 		PRURL       string `json:"pr_url"`
 	}
 	if err := decode(ev.Data, &p); err != nil {
